Add instructions for using turbo:load on ESBuild
Importmaps are the default way to use JS with Rails, but some apps (including those that use React or existed prior to Rails 7) don't use it yet. ESBuild is a very popular tool amongst Rails developers, and I feel it'd be useful to add the instructions on the official docs.